Move test directory to parent repo
