+++ test/parameters/constructClientParameterDetails.test.js
@@ -1,4 +1,4 @@
-import { constructClientParameterDetails } from '../../helpers/parameters/constructClientParameterDetails'
+import { constructClientParameterDetails } from '../../src/helpers/parameters/constructClientParameterDetails'
 
 describe('constructClientParameterDetails', function() {
   it('should properly return the value, title and placeholder for each parameter', function() {})
