+++ test/pullRequests/constructClientPullRequest.test.js
@@ -4,8 +4,8 @@ import chaiJestDiff from 'chai-jest-diff'
 
 chai.use(chaiJestDiff())
 
-import { constructPullRequestFromContractDetails } from '../../helpers/pullRequests/constructPullRequestFromContractDetails.js'
-import { FRONTEND_REPO_NAME } from '../../constants/repoNames'
+import { constructPullRequestFromContractDetails } from '../../src/helpers/pullRequests/constructPullRequestFromContractDetails.js'
+import { FRONTEND_REPO_NAME } from '../../src/constants/repoNames'
 
 describe('constructPullRequestFromContractDetails', function() {
   it('should properly construct client pullRequests', async function() {
