+++ test/tasks/decodeTaskBytes32ToMetaData.test.js
@@ -4,8 +4,8 @@ import chaiJestDiff from 'chai-jest-diff'
 
 chai.use(chaiJestDiff())
 
-import { decodeTaskBytes32ToMetaData } from '../../helpers/tasks/decodeTaskBytes32ToMetaData'
-import { getRepoNameFromNumber } from '../../helpers/tasks/getRepoNameFromNumber'
+import { decodeTaskBytes32ToMetaData } from '../../src/helpers/tasks/decodeTaskBytes32ToMetaData'
+import { getRepoNameFromNumber } from '../../src/helpers/tasks/getRepoNameFromNumber'
 
 describe('decodeTaskBytes32ToMetaData', function() {
   it('should properly decode a EVM task bytes32 into task metadata', function() {
