+++ test/tasks/encodeTaskBytes32ToMetaData.test.js
@@ -1,4 +1,4 @@
-import { encodeTaskMetaDataToBytes32 } from '../../helpers/tasks/encodeTaskMetaDataToBytes32'
+import { encodeTaskMetaDataToBytes32 } from '../../src/helpers/tasks/encodeTaskMetaDataToBytes32'
 
 describe('encodeTaskMetaDataToBytes32', function() {
   it('should properly encode task metadata into a EVM bytes32', function() {})
